Build: put model files into model folder and remove duplicate files that don't really do anything
